requirements.txt
gsplat==1.5.3
pycolmap==3.13.0
viser
nerfview==0.0.2
imageio[ffmpeg]
numpy<2.0.0
scikit-learn
tqdm
torchmetrics[image]
opencv-python
tyro>=0.8.8
Pillow
tensorboard
tensorly
pyyaml
matplotlib
kornia
easydict
plotly
evo==1.11.0
plyfile
protobuf
